+++ NEWS.md
@@ -1,3 +1,12 @@
+# cf 0.1.1
+
+Added ggplot2 contour functions.
+
+Added contour lines option.
+
+
 # cf 0.1.0
 
 First release.
+
+Accepted to CRAN on 5/4/17.